requirements.txt
python-jsonrpc
pymongo
pika
requests
redis
lxml
sklearn
python-dateutil
numpy
scipy
newspaper
